Still not passing part 2. Moving on
